+++ src/types/adapter-types.ts
@@ -1,19 +1,24 @@
 import { IDbRecord, IDbRecordBase, OmittedEntity } from './entity-types';
 import { DbSetCacheConfiguration, IDbSetInfo } from './dbset-types';
 import { EntitySelector } from './common-types';
+import { SearchResult } from '../common/SearchResult';
 
 export interface IDbSetFetchAdapter<TDocumentType extends string, TEntity extends IDbRecord<TDocumentType>, TExclusions extends keyof TEntity = never> {
+    filter(selector: (entity: TEntity, index?: number, array?: TEntity[]) => boolean): Promise<SearchResult<TDocumentType, TEntity, TExclusions>>;
+    find(selector: (entity: TEntity, index?: number, array?: TEntity[]) => boolean): Promise<SearchResult<TDocumentType, TEntity, TExclusions>>;
+    all(): Promise<SearchResult<TDocumentType, TEntity, TExclusions>>;
+    get(...ids: string[]): Promise<SearchResult<TDocumentType, TEntity, TExclusions>>;
+}
+
+export interface IDbSetFetchMediator<TDocumentType extends string, TEntity extends IDbRecord<TDocumentType>, TExclusions extends keyof TEntity = never> {
+    useCache(configuration: DbSetCacheConfiguration): void;
+    clearCache(...keys: string[]): void;
     filter(selector: (entity: TEntity, index?: number, array?: TEntity[]) => boolean): Promise<TEntity[]>;
     find(selector: (entity: TEntity, index?: number, array?: TEntity[]) => boolean): Promise<TEntity | undefined>;
     first(): Promise<TEntity | undefined>;
     all(): Promise<TEntity[]>;
     get(...ids: string[]): Promise<TEntity[]>;
     pluck<TKey extends keyof TEntity>(selector: EntitySelector<TDocumentType, TEntity>, propertySelector: TKey): Promise<TEntity[TKey]>
-}
-
-export interface IDbSetFetchMediator<TDocumentType extends string, TEntity extends IDbRecord<TDocumentType>, TExclusions extends keyof TEntity = never> extends IDbSetFetchAdapter<TDocumentType, TEntity, TExclusions> {
-    useCache(configuration: DbSetCacheConfiguration): void;
-    clearCache(...keys: string[]): void;
 }
 
 export interface IDbSetGeneralAdapter<TDocumentType extends string, TEntity extends IDbRecord<TDocumentType>, TExclusions extends keyof TEntity = never> {